+++ src/main/java/com/example/pbl3_1/mapper/ProductItemMapper.java
@@ -14,7 +14,7 @@ public class ProductItemMapper implements RowMapper<ProductItem> {
             pi.setVariation1(rs.getLong("variation1"));
             pi.setVariation2(rs.getLong("variation2"));
             pi.setQtyInStock(rs.getInt("qty_in_stock"));
-            pi.setPrice(rs.getFloat("price"));
+            pi.setPrice(rs.getInt("price"));
             pi.setProductImgPath(rs.getString("product_img_path"));
             pi.setProductId(rs.getLong("product_id"));
         } catch (Exception e) {
